pkg/sequence/parser/ast: moves the product of parser into the parser package
