+++ pkg/sequence/parser/match/match.go
@@ -3,7 +3,7 @@ package match
 import (
 	"regexp"
 
-	"github.com/ufukty/diagramer/pkg/sequence/ast"
+	"github.com/ufukty/diagramer/pkg/sequence/parser/ast"
 )
 
 var (
